Begin work to allow apps to specify which database instances they want to depend on.
Rather than specify the instances themselves, I plan for the apps to declare what kind of support they need from the cluster. For example, read only, active master, bulkhead etc. The cluster can then decide which of its instances best serves the requirement. The immediate requirement for this is to be able to (manually) switch the active master, and also select a particular database instance to use as a bulkhead for a certain group of apps.

So far just a sketch of the first test is implemented, which passes with no change in behaviour, just enough code to show the API.