Orc resolve after reprovision (again)
The orc resolve after reprovision was lost when the reprovision task stopped
depending on the provision task and instead on the provision_machine task. The
provision task called orc:resolve in its body. It did it this way because it
was conditional on the orc:resolve task actually existing. However the
orc:resolve task would only every exist when there are no applications to
resolve in the stack. In this case the resolve can still exist, but will just
be a noop. This change introduces orc:resolve for everything so that we can
just use regular rake dependencies and not have to complicate things with
conditional calls.