Handle another unexpected layer of what a machine_def can be.
A so-called "machine_def" may not only be a MachineDef or a MachineDefContainer, it can also be a CustomServices. This change makes the dependencies task able to traverse these three layers in order to produce a correct list of fqdns regardless of what level the machine_def var is at.