mpimm: Allow stacks to depend on standalone_app_services
Required so that residual returns apps can depend on the file feed cache
standalone app.

This avoids these dependencies having to be declared with allow_host.

There is still no way to generate the dependencies (config url's) in the
dependant app. If there were more than one instance of the the
standalone app we wouldn't know which one to depend on.