GrahamA: members of a MySQL cluster are now modelled as depending on each other.
This means that provisioning a new database server will prepare all the other members of the cluster (giving e.g. mysql accounts, firewall rules, etc). The most obvious use case for this is provisioning a new replica, which previously would not be able to start replication without preparing it's primary server with a puppet run.