Fix a problem I had making clean_host work.
I believe this was incorrectly moved outside of the Proc in commit
fffbc8bd688571aff8c6bf011b244416724e98bao

It causes problems when trying to migrate a VM with no data storage if it is
considered during allocation after a VM with data storage. The current VM being
considered 'inherited' the data storage size from the previous VM and
allocation would fail.