Provide participation control
The participation label gives control over whether a pod is part of the
service or not. By changing the label on a pod to
`participation=disabled` the deployment will start a new pod and the
service will removed the relabeled pod from the service. Likewise, by
relabeling to `participation=enabled` the pod will rejoin the deployment
and the service. The deployment will also rescale to have the
appropriate number of replicas.