Split service-external and service-internal deps
The dependency_nodes code confused the idea of dependencies by combining
dependencies between services and dependencies inside a service. The
dependencies between services (service-external) information is tracked
by the dependencies.rb code and is only between "Virtual Services". The
dependencies between instances inside a service (service-internal) was
tacked on by the dependency nodes and caused dependencies to be between
individual "Machine Defs".

In the world of kubernetes this becomes a problem because different
aspects of dependencies were suddenly working with different levels of
the model.

This tries to fix it by divorcing the two parts and letting the
consumers of dependencies put things back together in ways that make
sense for them.