Stop duplicate ipBlock entries.
endpoints returns the same as other code, which is an array of mysql
hostnames, where the first entry is the server you can write to (often
the master). In certain circumstances you also get the same hostname as
the second entry. From the point of view of network policies we only
care that we have a unique list of IP addresses.