Pass the config file through an initContainer
The config file will actually be a template that needs to be filled in
with secrets. To do that we are going to use an initContainer that will
construct the final config.properties file.

This is the first step of that and has the config file copied into place
by an initContainer. After this, the initContainer will get environment
variables for the secrets it needs and will interpolate those into the
place.

For now I'd like to keep the script to do this interpolation in stacks  as a
way of clearly versioning it with stacks.