Only output the resources in compile
The introduction of the KubernetesResources class caused it to show up
serialized in the compile output. This happened because it was
Enumerable. The decision to make it Enumerable was made because it would
limit the changes to the rest of the code and the most common use case
was to iterate across the resources. However, that introduced the easy
possibility of passing it along when it wasn't intended.

Combine outputing it to YAML with the fact that deserializing YAML
recreates the class, which is available in the tests, it means that the
test would need to assert the type of deserialized YAML (ick). By just
removing the Enumerable and introducing an accessor for the k8s
resources it causes each user to explicitly select the resources and
uncovers the error.