Fix secret lookups.
These happen on the puppet server via the mcollective agent, where we
use 'logicalenv' to refer to the stacks environment name. We do this
because we use puppets environment concept to use run different puppet
branches.

Slightly confusing to have @logicalenv in app_service when @environment
is as thing, BUT you're just writing the template for the config file.
When the template is populated it's actually done somewhere else.