Control memory resources
Memory limits and requests are based on the jvm heap size that the
application is configured for. The default is to provide 10% of the heap
size as headspace for non-heap jvm use + other process use (such as an
exec in the container). The heap size is configured with jvm_heap and
the extra space percentage is configured with headspace.

For example the following configuration:
  self.jvm_heap = '250M'
  self.headspace = 0.5

will result in the JVM being started with -Xmx250M and the container
getting resources limits and requests of 375Mi (= 250 * (1 + 0.5)). Note
that M base-10 for kubernetes, but base-2 for the JVM. Mi (mebibytes),
is the base-2 name.