Consistently name hiera keys for mysql passwords
The password hiera key should be <environment>/<app
name>/mysql_passwords. The short name shouldn't be used in this name at
all. This is even more complicated by the fact that the "standard
service", which is really for non-standard services, doesn't have an
application name. To fix this, the change was made to use a "database
application name", which each service type can determine what it is.