Fix app.kubernetes.io/instance label.
Using the name of the app for the last part makes no sense as it limits
the use of an application to once application per stack. But we imagine
that we oculd have the same application multiple times per stack but
with a difference service name. So use the servie name instead.