Include rights even without fqdn dependants
The code assumed that it was only worthwhile including application
rights if there are fqdns of systems depending on the database. This
isn't true if only kubernetes services depend on the database.

The guard to check for the fqdns is only interesting if including an
empty hash for the application_rights_wrapper would have an adverse
effect. That puppet code just calls create_resources on the hash, so an
empty hash will simply have no effect.