Alex/Andy: Support multi-site kubernetes and same_site dependency
The kubernetes services now support being placed in multiple sites using
the instances hash form. In addition kubernetes network policies
understand a requirement of :same_site to provide network policies only
for services that live in the same site as the dependent service.

This is put in place for the insight dashboard to work. There are
probably corner cases that aren't covered correctly for general use.