Set the min heap size as well as the max
Leaving off the min size means that the JVM will slowly grow the heap to
the max as memory is used. However before it grows the heap it will
perform a full GC. This means that it spends a lot of time on full GCs
as the system "warms up" to the full size. As the heap grows it performs
full GCs less and less.

Our VM systems set Xms and this was lost in the translation to k8s.