Andy/Alex: be more lenient about taking pods out of services
Previously we were taking the default settings of 1 second timeout, 2
second period and 3 failureThreshold. With these settings we saw pods
being taken out of the service after a GC pause of 6 seconds. With these
new settings we are aiming to replicate as closely as possible the
exising loadbalancer/keepalived settings which have a 10 second timeout
and (implicit) failureThreshold of 1.

We want multiple failures before we remove the pod from the service so
that we can create an alert based on readinessProbe failures to notify
maintainers.