K8s internal deps go through services
When one k8s app depends on another k8s app it was given the external
VIP/DNS name for the app. However, that route isn't available from within
the the cluster. The app should be going through the app's kubernetes
Service so that it gets the load balancing it needs and with the fewest
hops.