Express the VM model as Prometheus metrics
In order to search for missing VM datapoints we need to have a list of
what should be sending data. This provides that list and a whole lot
more.

To get this into the clusters we'll setup a jenkins job to run whenever
the config changes. That job will produce the resource and pass it to
kubectl apply