Support declaration of :all dependency
The insight dashboard needs to depend on everything in the same site as
it. The current way it does this means that it needs to be reprovisioned
every time a new app is deployed or else you can't see the status pages
of the app. This creates a stumbling block for an activity that will
become more and more common as things migrate.

In order to remove this problem, this commit implements the ability to
declare: depend_on :all, :all, :same_site. That will be translated to a
dependency using a pod selector of anything managed by stacks rather
than a selector for pods of a specific application.