First pass at proxy server proxy_pass_rules supporting multiple services.
This would allow a proxy pass rules 'service' parameter to take an array of services to hand off to rather than just the name of a single service.
The proxy and dependency code then uses the 'index' of the machine_def to govern which of those services a particular proxy hands off to.

Outstanding things to fix/complete are:
- Reverse dependencies, for example right now a proxy that adds proxy pass rules to work like this no longers has firewall rules to allow it to talk to the actual application server (this is caused by the return added to dependencies.rb in this commit)
- Checking that the correct NetworkPolicies are created on the k8s services, allowing the correct proxy access to them.