Show what stack failed to compile
When compiling everything it is hard to tell what had the error. By
reporting that we can more easily debug. In addition, the stack trace is
more of debugging information that anything else. By hiding it in DEBUG,
we can make the error a little more evident to the user.