Remove the ability to specify the certificate
Specifying the certificate in stacks is redundant and a slow process for
rolling out a new certificate. The puppet code already has the same
defaults defined and can perform similar logic needed (changing by
environment) when performing a rollout. Having it in here simply means
that it takes longer to roll out any changes and duplicates the default,
which now needs to be updated in more places.