aparker/mpimm: Prepare for removing all port names other than app or metrics
In particular for ntp server and cieye that use ntp and http. We had a
selection of different port names, but no consistent way to use them. By
limiting the number of names (app, metrics) each port now has a specific
purpose that we can code to. App is for any application specific traffic
(such as the applications user requests) and metrics is for any
prometheus metric traffic. In the future if there is need for more types
of traffic, we can add more ports.