Allow only app and metrics as ports
This is the final move to limit the ports that can be defined for
stacks. The purpose of the limitation is so that each port can be
specifically wired through the system in the correct way. The 'app' port
is for application specific traffic. The 'metrics' port is for
prometheus metrics to be exposed. The 'metrics' port is only ever
exposed to prometheus whereas the 'app' port is exposed to other
applications inside the cluster or via an ingress to systems outside the
cluster (when a 'service_port' is given).