Redact usernames that can show up in TIM requests
The TIM has an old system for username and password login that puts
j_username and j_password into the query string. This causes the
password to show up in any access logs that a server produces. To get
around this in Apache we piped all logs through a sed command that
removes the password. This duplicates the same command for traefik.