Expose metrics on the application's service
The attempt to put all metrics on a common service caused a problem. The
TargetDown alert became uninformative as it didn't tell us which target
was having problem. Built into that alert is the thought that a service
is a particular application and that alerting on that service being down
is meaningful. The metrics service that I created broke all of those
assumptions. Therefore putting the metrics on the application's service
seems like the best route.