Only one networkpolicy per service
When a service (like rabbitmq) is depended on for two different reasons
we created two networkpolicies. The network policies however were the
exact same.

Also the hash of the policy flipped depending on the order in which a
hash or uniq returned the items. By sorting we get rid of that problem.