Andy/Andrew: Start & Stop apps in dep order
The nightly refresh of latest-timdb caused an error from the TIM every
night because abacus was shut down first. Based on the dependencies
between the apps in stacks, Abacus needs to be shut down after TIM. This
change does a simple (very simple) topological sort of the dependents of
the target (TIM DB in the real case) so that that the used service is
shut down after the using service.